Ready to use js specs
